Implement major game logic.
This commit was mostly concerned with the game logic and javascript,
although I did make a few layout tweaks as well as add a few different elements for displaying information and starting a new game.

The two main parts of the game logic are the gameBoard module and the displayController module.
The gameBoard module holds the status of the game board in an array. The array cannot be accessed directly, but through other functions can be manipulated.
The gameBoard module includes a few different methods for making a move, checking if the board is full, getting the board array, and reseting the board array.
It's important to note that there are not many restrictions within the move method. All of the checks and other validation are done via the displayController.

The displayController contains many different methods for controlling the flow of the game. Only two are publicly accessible: addMark() and startGame(), as these are the only methods where the user must take some sort of action.
The displayController stores information about the game status, such as the number of moves made, whether or not the game is over, and the two players' information.
The displayController is in charge of filling in the board in the DOM with the gameBoard information, allowing the users to make new moves, checking if the game is over, and handling the end game flow.